+++ Realm/Realm/Realm.cs
@@ -175,6 +175,7 @@ namespace Realms
 
         private State _state;
         private WeakReference<Session> _sessionRef;
+        private Transaction _activeTransaction;
 
         internal readonly SharedRealmHandle SharedRealmHandle;
         internal readonly RealmMetadata Metadata;
@@ -427,6 +428,8 @@ namespace Realms
         {
             if (!IsClosed)
             {
+                _activeTransaction?.Dispose();
+
                 if (SharedRealmHandle.OwnsNativeRealm)
                 {
                     _state.RemoveRealm(this);
@@ -767,7 +770,9 @@ namespace Realms
             ThrowIfFrozen("Starting a write transaction on a frozen Realm is not allowed.");
 
             SharedRealmHandle.BeginTransaction();
-            return new Transaction(this);
+
+            Debug.Assert(_activeTransaction == null, "There should be no active transaction");
+            return _activeTransaction = new Transaction(this);
         }
 
         /// <summary>
@@ -885,7 +890,9 @@ namespace Realms
             }
 
             await SharedRealmHandle.BeginTransactionAsync(synchronizationContext, cancellationToken);
-            return new Transaction(this);
+
+            Debug.Assert(_activeTransaction == null, "There should be no active transaction");
+            return _activeTransaction = new Transaction(this);
         }
 
         /// <summary>
@@ -1595,6 +1602,7 @@ namespace Realms
 
         internal void DrainTransactionQueue()
         {
+            _activeTransaction = null;
             _state.DrainTransactionQueue();
         }